+++ Changelog.md
@@ -1,5 +1,6 @@
 # v0.8.4 2015-09-xx
 
+* Add mutation from `a != b` to `!a.eql?(b)` and `!a.equal?(b)` #417
 * Add mutation `A.const_get(:B)` -> `A::B` #426
 
 # v0.8.3 2015-09-01
@@ -7,7 +8,6 @@
 * Remove invalid mutation `super(...)` to `super`
 * Add mutation from `def foo(a = true); end` to `def foo(a = true); a = true; end` #419
 * Add mutation from `def foo; end` to `remove_method :foo` #413
-* Add mutation from `a != b` to `!a.eql?(b)` and `!a.equal?(b)` #417
 
 # v0.8.2 2015-08-11